Removes class BBBMeeting from Apidoc, modifies usage of new Date
